Separate visual and app layer in vouchers
