+++ src/discounts/components/VoucherListPage/index.ts
@@ -1,2 +1,3 @@
 export { default } from "./VoucherListPage";
 export * from "./VoucherListPage";
+export * from "./filters";
