--- src/discounts/views/VoucherList/messages.ts
@@ -1,41 +1,0 @@
-import { defineMessages } from "react-intl";
-
-const messages = defineMessages({
-  active: {
-    defaultMessage: "Active",
-    description: "voucher status"
-  },
-  expired: {
-    defaultMessage: "Expired",
-    description: "voucher status"
-  },
-  fixed: {
-    defaultMessage: "Fixed amount",
-    description: "discount type"
-  },
-  percentage: {
-    defaultMessage: "Percentage",
-    description: "discount type"
-  },
-  scheduled: {
-    defaultMessage: "Scheduled",
-    description: "voucher status"
-  },
-  started: {
-    defaultMessage: "Started",
-    description: "voucher start date"
-  },
-  status: {
-    defaultMessage: "Status",
-    description: "voucher status"
-  },
-  timesUsed: {
-    defaultMessage: "Times used",
-    description: "voucher"
-  },
-  type: {
-    defaultMessage: "Discount Type"
-  }
-});
-
-export default messages;
